# Vectorize correlation computation