Based on discussion with @nx10 and @amygutierrez, it seems there might be room to speed up regression testing significantly. This PR makes an initial attempt by vectorizing the computation of Pearson and concordance correlation coefficients.

I checked that the new `batch_correlate()` and old `correlate()` give the same result on random-normal data, up to numerical precision.

I also benchmarked `batch_correlate()` against the old `correlate_two_nifti_timeseries()`:

```python
x = np.random.randn(100, 100, 100, 100)
y = np.random.randn(100, 100, 100, 100)

# 2min 46s ± 1.47 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
%timeit correlate_two_nifti_timeseries(x, y, x.shape)

# 4.5 s ± 85.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
%timeit batch_correlate(x, y)
```

The speedup is ~36x for data of this size on my laptop.

## Checklist
<!-- Replace  the [ ] with [x] to check the boxes. -->
- [x] My pull request has a descriptive title (not a vague title like `Update index.md`).
- [x] My commit messages follow best practices.
- [x] My code follows the established code style of the repository.
- [ ] I added tests for the changes I made (if applicable).
- [ ] I updated the changelog.
- [ ] I added or updated documentation (if applicable).
- [x] I tried running the project locally and verified that there are no visible errors.

## Developer Certificate of Origin
<!-- You must read and understand the following attestation. -->

<details>
<summary>Developer Certificate of Origin</summary>

```
Developer Certificate of Origin
Version 1.1

Copyright (C) 2004, 2006 The Linux Foundation and its contributors.
1 Letterman Drive
Suite D4700
San Francisco, CA, 94129

Everyone is permitted to copy and distribute verbatim copies of this
license document, but changing it is not allowed.


Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.
```

</details>